requirements.txt
fastapi==0.87.0
loguru==0.6.0
pydantic==1.10.2
uvicorn==0.20.0
PyYAML==6.0.1
PyJWT==2.6.0
base58==2.1.1
py-near==1.1.31